Fix the `run` subcommand, ditch overlayfs
Overlayfs was the best documented working way to have a readonly bind-mount.
I really thought there were no other.
It worked fine until it stopped working for a mysterious external reason.
Years passed.
Now that I'm no longer alone to experience this, symptoms point at the
disabling of overlapping layers in Overlayfs (linux commit 146d62e5a5867).
Duh, that was the point! Every directory overlaps with the root directory,
and I still want a readonly mirror of it!

In the meantime, mount(1) has gained support for readonly bind-mounts that
actually are that, instead of unhelpfully remounting the source readonly,
demonstrating that such a thing now exists.

The secret is in the mount(2) manpage: A readonly bind-mount can be had by
remounting with all of the "remount,bind,ro" options set.

Fixes issue #1.